Organise files around components instead of roles. Inspired from https://blog.risingstack.com/node-hero-node-js-project-structure-tutorial/
