+++ index.route.js
@@ -1,8 +1,10 @@
 import express from 'express';
-import userRoutes from './user.route';
-import authRoutes from './auth.route';
+import userRoutes from './server/user/user.route';
+import authRoutes from './server/auth/auth.route';
 
 const router = express.Router(); // eslint-disable-line new-cap
+
+// TODO: use glob to match *.route files
 
 /** GET /health-check - Check service health */
 router.get('/health-check', (req, res) =>
