+++ server/auth/auth.route.js
@@ -2,7 +2,7 @@ import express from 'express';
 import validate from 'express-validation';
 import expressJwt from 'express-jwt';
 import paramValidation from '../../config/param-validation';
-import authCtrl from '../controllers/auth.controller';
+import authCtrl from './auth.controller';
 import config from '../../config/config';
 
 const router = express.Router(); // eslint-disable-line new-cap
